pyproject.toml
[tool.poetry]
name = "word-searcher"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
psycopg2 = "^2.9.10"
python-dotenv = "^1.0.1"
bs4 = "^0.0.2"
pandas = "^2.2.3"
sqlalchemy = "^2.0.38"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
